requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.3
autopep8==1.5
bcrypt==3.1.6
blinker==1.4
Brlapi==0.6.7
cachetools==4.0.0
cbpro==1.1.4
certifi==2018.8.24
chardet==3.0.4
coinbase==2.1.0
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.12
defer==1.0.6
devscripts===2.19.6ubuntu1
distro==1.3.0
distro-info===0.21ubuntu4
duplicity==0.8.4
entrypoints==0.3
fasteners==0.12.0
future==0.16.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gpg==1.12.0
httplib2==0.11.3
idna==2.6
isort==4.3.21
keyring==18.0.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.10.0
macaroonbakery==1.2.3
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.1.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycryptodome==3.9.7
pycups==1.9.73
PyGObject==3.34.0
PyJWT==1.7.0
pylint==2.4.4
pymacaroons==0.13.0
pymongo==3.5.1
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.7.3
python-debian==0.1.36
python-magic==0.4.16
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
reportlab==3.5.23
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.1.0
system-service==0.3
systemd-python==234
typed-ast==1.4.1
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
uritemplate==3.0.1
urllib3==1.24.1
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.40.0
wrapt==1.11.2
xkit==0.0.0
